Command/Inst refactor: Pulses and Play (#3936)
* Deprecate PulseCommands, create new Play Instruction-implementation. The Play instruction takes a Pulse and a PulseChannel. Pulse is a new class within the pulse library that is very close to the deprecated PulseCommand abstract class. The PulseCommand implementations (SamplePulse and the ParametricPulse subclasses) have been moved to the pulse library as well, subclassing from Pulse, and all valid arguments to Play.

* Update import statements in Pulse

* Fill in the implementation of some IP work like __call__

* Fill in and update documentation, complete import paths

* fixup cyclic imports

* IP updates to assemble_schedules for supporting Play instruction

* Give all pulses a name arg and update reno note

* Continue pulse_instruction conversion support for Play

* Add support for Play in assemble schedules, fix style

* Attempt to fix build error

* Add missing name field to Constant pulse

* To support __call__ from Pulse, I need to import Play, which means I can't import Pulse from Play (for typehints) for the timebeing

* Fixup bugs introduced in SamplePulse

* Update tests with new API

* Add tests and fixup missing name passing

* Fixup tests, remove deprecation warning from assemble execution, add hash and eq methods to pulses

* Fixup docs

* Update functional_pulse import path and Instruction type ref

* try again with the type hints

* Was missing updates to PulseCommand that needed to be migrated with the new Pulse class

* Was missing docstring type documentation for unresolvable type

* Was missing removal of PulseStyle in parametric pulses

* The changes from the sphinx warrnings pass hadn't been moved with the migrated ParametricPulses module. Moved the drag pulse docstring over finally

* Fix spacing in drag pulse docstring

* The docs should finally build now

* Update qiskit/pulse/instructions/play.py


* Update qiskit/pulse/pulse_lib/sample_pulse.py


* Apply suggestions from code review


* Documentation improvements

* Test change because linux python35 is failing

* Update releasenotes/notes/unify-instructions-and-commands-aaa6d8724b8a29d3.yaml


* Fix plotting Play instruction

* style, line length

* Attempt to fix failing test by comparing names directly

* Handle all instruction types in add_instruction

* Remove unused import

* Update releasenotes/notes/unify-instructions-and-commands-aaa6d8724b8a29d3.yaml


* Separate note about Play feature into two notes: Play feature and Pulse movement
