+++ qiskit/visualization/pulse/matplotlib.py
@@ -33,7 +33,7 @@ from qiskit.pulse.channels import (DriveChannel, ControlChannel,
                                    MeasureChannel, AcquireChannel,
                                    SnapshotChannel)
 from qiskit.pulse.commands import FrameChangeInstruction
-from qiskit.pulse import (SamplePulse, FrameChange, PersistentValue, Snapshot,
+from qiskit.pulse import (SamplePulse, FrameChange, PersistentValue, Snapshot, Play,
                           Acquire, PulseError, ParametricPulse, SetFrequency, ShiftPhase)
 
 
@@ -59,18 +59,23 @@ class EventsOutputChannels:
         self._labels = None
         self.enable = False
 
-    def add_instruction(self, start_time, pulse):
+    def add_instruction(self, start_time, instruction):
         """Add new pulse instruction to channel.
 
         Args:
             start_time (int): Starting time of instruction
-            pulse (Instruction): Instruction object to be added
+            instruction (Instruction): Instruction object to be added
         """
-
-        if start_time in self.pulses.keys():
-            self.pulses[start_time].append(pulse.command)
+        if instruction.command is not None:
+            pulse = instruction.command
+        elif isinstance(instruction, Play):
+            pulse = instruction.pulse
         else:
-            self.pulses[start_time] = [pulse.command]
+            pulse = instruction
+        if start_time in self.pulses.keys():
+            self.pulses[start_time].append(pulse)
+        else:
+            self.pulses[start_time] = [pulse]
 
     @property
     def waveform(self):
