+++ releasenotes/notes/unify-instructions-and-commands-aaa6d8724b8a29d3.yaml
@@ -11,12 +11,14 @@ features:
 
         sched += Delay(5)(DriveChannel(0))
         sched += ShiftPhase(np.pi)(DriveChannel(0))
+        sched += SamplePulse([1.0, ...])(DriveChannel(0))
         sched += Acquire(100)(AcquireChannel(0), MemorySlot(0))
 
     or, equivalently (though less used)::
 
         sched += DelayInstruction(Delay(5), DriveChannel(0))
         sched += ShiftPhaseInstruction(ShiftPhase(np.pi), DriveChannel(0))
+        sched += PulseInstruction(SamplePulse([1.0, ...]), DriveChannel(0))
         sched += AcquireInstruction(Acquire(100), AcquireChannel(0),
                                     MemorySlot(0))
 
@@ -25,10 +27,26 @@ features:
 
         sched += Delay(5, DriveChannel(0))
         sched += ShiftPhase(np.pi, DriveChannel(0))
-        sched += Acquire(100, AcquireChannel(0), MemorySlot(0))
+        sched += Play(SamplePulse([1.0, ...]), DriveChannel(0))
         sched += SetFrequency(5.5, DriveChannel(0))  # New instruction!
+        sched += Acquire(100, AcquireChannel(0), MemorySlot(0))
+  - |
+    There is now a :py:class:`~qiskit.pulse.instructions.Play` instruction
+    which takes a description of a pulse envelope and a channel. There is a
+    new :py:class:`~qiskit.pulse.pulse_lib.Pulse` class in the ``pulse_lib``
+    from which the pulse envelope description should subclass.
 
+    For example::
+
+        Play(SamplePulse([0.1]*10), DriveChannel(0))
+        Play(ConstantPulse(duration=10, amp=0.1), DriveChannel(0))
 deprecations:
+  - |
+    py:class:`~qiskit.pulse.pulse_lib.SamplePulse` and
+    py:class:`~qiskit.pulse.pulse_lib.ParametricPulse` s (e.g. ``Gaussian``)
+    now subclass from :py:class:`~qiskit.pulse.pulse_lib.Pulse` and have been
+    moved to the ``pulse_lib``. The previous path via ``pulse.commands`` is
+    deprecated.
   - |
     ``DelayInstruction`` has been deprecated and replaced by
     :py:class:`~qiskit.pulse.instruction.Delay`. This new instruction has been
@@ -56,6 +74,12 @@ deprecations:
     calling a command on a channel will be supported::
 
         ShiftPhase(<phase>)(<channel>)
+  - |
+    The ``call`` method of py:class:`~qiskit.pulse.pulse_lib.SamplePulse` and
+    py:class:`~qiskit.pulse.pulse_lib.ParametricPulse` s have been deprecated.
+    The migration is as follows::
+
+        Pulse(<*args>)(<channel>) -> Play(Pulse(*args), <channel>)
   - |
     ``AcquireInstruction`` has been deprecated and replaced by
     :py:class:`~qiskit.pulse.instructions.Acquire`. The changes are::
