+++ test/python/qobj/test_pulse_converter.py
@@ -206,14 +206,12 @@ class TestQobjToInstructionConverter(QiskitTestCase):
 
     def test_drive_instruction(self):
         """Test converted qobj from PulseInstruction."""
-        cmd = self.linear
-        instruction = cmd(DriveChannel(0)) << 10
+        instruction = self.linear(DriveChannel(0))
 
         qobj = PulseQobjInstruction(name='linear', ch='d0', t0=10)
         converted_instruction = self.converter(qobj)
 
-        self.assertEqual(converted_instruction.timeslots, instruction.timeslots)
-        self.assertEqual(converted_instruction.instructions[0][-1].command, cmd)
+        self.assertEqual(converted_instruction.instructions[0][-1], instruction)
 
     def test_parametric_pulses(self):
         """Test converted qobj from ParametricInstruction."""
@@ -226,7 +224,7 @@ class TestQobjToInstructionConverter(QiskitTestCase):
             parameters={'duration': 25, 'sigma': 15, 'amp': -0.5 + 0.2j})
         converted_instruction = self.converter(qobj)
         self.assertEqual(converted_instruction.timeslots, instruction.timeslots)
-        self.assertEqual(converted_instruction.instructions[0][-1].command, instruction.command)
+        self.assertEqual(converted_instruction.instructions[0][-1], instruction)
 
     def test_frame_change(self):
         """Test converted qobj from FrameChangeInstruction."""
