+++ test/python/visualization/test_pulse_visualization_output.py
@@ -23,9 +23,8 @@ from qiskit.tools.visualization import HAS_MATPLOTLIB
 from qiskit.visualization import pulse_drawer
 from qiskit.pulse.channels import (DriveChannel, MeasureChannel, ControlChannel, AcquireChannel,
                                    MemorySlot, RegisterSlot)
-from qiskit.pulse.commands import (FrameChange, Acquire, ConstantPulse, Snapshot, Delay,
-                                   Gaussian)
-from qiskit.pulse.instructions import SetFrequency
+from qiskit.pulse.commands import FrameChange
+from qiskit.pulse.instructions import SetFrequency, Play, Acquire, Delay, Snapshot
 from qiskit.pulse.schedule import Schedule
 from qiskit.pulse import pulse_lib
 
@@ -65,7 +64,8 @@ class TestPulseVisualizationImplementation(QiskitVisualizationTestCase):
         delay = Delay(100)
         sched = Schedule()
         sched = sched.append(gp0(DriveChannel(0)))
-        sched = sched.insert(0, ConstantPulse(duration=60, amp=0.2 + 0.4j)(ControlChannel(0)))
+        sched = sched.insert(0, pulse_lib.ConstantPulse(duration=60, amp=0.2 + 0.4j)(
+            ControlChannel(0)))
         sched = sched.insert(60, FrameChange(phase=-1.57)(DriveChannel(0)))
         sched = sched.insert(60, SetFrequency(8.0, DriveChannel(0)))
         sched = sched.insert(30, gp1(DriveChannel(1)))
@@ -85,9 +85,22 @@ class TestPulseVisualizationImplementation(QiskitVisualizationTestCase):
     @unittest.skip('Useful for refactoring purposes, skipping by default.')
     def test_parametric_pulse_schedule(self):
         """Test that parametric instructions/schedules can be drawn."""
-        filename = self._get_resource_path('current_schedule_matplotlib_ref.png')
+        filename = self._get_resource_path('current_parametric_matplotlib_ref.png')
         schedule = Schedule(name='test_parametric')
-        schedule += Gaussian(duration=25, sigma=4, amp=0.5j)(DriveChannel(0))
+        schedule += pulse_lib.Gaussian(duration=25, sigma=4, amp=0.5j)(DriveChannel(0))
+        pulse_drawer(schedule, filename=filename)
+        self.assertImagesAreEqual(filename, self.parametric_matplotlib_reference)
+        os.remove(filename)
+
+    @unittest.skipIf(not HAS_MATPLOTLIB, 'matplotlib not available.')
+    @unittest.skip('Useful for refactoring purposes, skipping by default.')
+    def test_play(self):
+        """Test that Play instructions can be drawn. The output should be the same as the
+        parametric_pulse_schedule test.
+        """
+        filename = self._get_resource_path('current_play_matplotlib_ref.png')
+        schedule = Schedule(name='test_parametric')
+        schedule += Play(pulse_lib.Gaussian(duration=25, sigma=4, amp=0.5j), DriveChannel(0))
         pulse_drawer(schedule, filename=filename)
         self.assertImagesAreEqual(filename, self.parametric_matplotlib_reference)
         os.remove(filename)
